refactor(web): route problem-sets pages through actions/queries
