+++ web/src/actions/problem-sets/index.ts
@@ -1,3 +1,4 @@
 export * from "./crud";
 export * from "./items";
 export * from "./likes";
+export * from "./queries";
